Rename branch-and-cut to branch-and-bound
